requirements.txt
google.cloud
google.cloud.language
flask